+++ src/apps/backend/services/gateway/modules/proxy/proxy.module.ts
@@ -1,0 +1,9 @@
+import { Module } from "@nestjs/common";
+
+import { ProxyController } from "./presentation/proxy.controller";
+
+
+@Module({
+    controllers: [ProxyController],
+})
+export class ProxyModule {}