+++ src/apps/backend/services/service-1/modules/presentation.module.ts
@@ -1,0 +1,9 @@
+import { Module } from "@nestjs/common";
+
+import { PresentationController } from "./presentation/presentation.controller";
+
+
+@Module({
+    controllers: [PresentationController],
+})
+export class PresentationModule {}