+++ src/apps/backend/services/service-1/modules/presentation/presentation.controller.ts
@@ -1,0 +1,17 @@
+import { Controller } from "@nestjs/common";
+import { MessagePattern } from "@nestjs/microservices";
+
+
+@Controller()
+export class PresentationController {
+    @MessagePattern({ cmd: "check" })
+    check() {
+        const params = {
+            service: "service-1",
+            status: "ok",
+            timestamp: new Date().toISOString(),
+        };
+
+        return params;
+    }
+}