+++ src/apps/backend/services/service-2/app.module.ts
@@ -1,0 +1,9 @@
+import { Module } from "@nestjs/common";
+
+import { PresentationModule } from "./modules/presentation.module";
+
+
+@Module({
+    imports: [PresentationModule],
+})
+export class AppModule {}